% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{associate_tag_option_with_resource}
\alias{associate_tag_option_with_resource}
\title{Associate the specified TagOption with the specified portfolio or product}
\usage{
associate_tag_option_with_resource(ResourceId, TagOptionId)
}
\arguments{
\item{ResourceId}{The resource identifier.}

\item{TagOptionId}{The TagOption identifier.}
}
\description{
Associate the specified TagOption with the specified portfolio or product.
}
\examples{

}
