% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{create_portfolio}
\alias{create_portfolio}
\title{Creates a portfolio}
\usage{
create_portfolio(DisplayName, ProviderName, IdempotencyToken,
  AcceptLanguage = NULL, Description = NULL, Tags = NULL)
}
\arguments{
\item{DisplayName}{The name to use for display purposes.}

\item{ProviderName}{The name of the portfolio provider.}

\item{IdempotencyToken}{A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Description}{The description of the portfolio.}

\item{Tags}{One or more tags.}
}
\description{
Creates a portfolio.
}
\examples{

}
