% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{create_portfolio_share}
\alias{create_portfolio_share}
\title{Shares the specified portfolio with the specified account}
\usage{
create_portfolio_share(PortfolioId, AccountId, AcceptLanguage = NULL)
}
\arguments{
\item{PortfolioId}{The portfolio identifier.}

\item{AccountId}{The AWS account ID.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Shares the specified portfolio with the specified account.
}
\examples{

}
