% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{create_tag_option}
\alias{create_tag_option}
\title{Creates a TagOption}
\usage{
create_tag_option(Key, Value)
}
\arguments{
\item{Key}{[required] The TagOption key.}

\item{Value}{[required] The TagOption value.}
}
\description{
Creates a TagOption.
}
\section{Accepted Parameters}{
\preformatted{create_tag_option(
  Key = "string",
  Value = "string"
)
}
}

