% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{delete_provisioned_product_plan}
\alias{delete_provisioned_product_plan}
\title{Deletes the specified plan}
\usage{
delete_provisioned_product_plan(PlanId, AcceptLanguage = NULL,
  IgnoreErrors = NULL)
}
\arguments{
\item{PlanId}{The plan identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{IgnoreErrors}{If set to true, AWS Service Catalog stops managing the specified provisioned product even if it cannot delete the underlying resources.}
}
\description{
Deletes the specified plan.
}
\examples{

}
