% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{delete_provisioning_artifact}
\alias{delete_provisioning_artifact}
\title{Deletes the specified provisioning artifact (also known as a version) for the specified product}
\usage{
delete_provisioning_artifact(AcceptLanguage = NULL, ProductId,
  ProvisioningArtifactId)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProductId}{[required] The product identifier.}

\item{ProvisioningArtifactId}{[required] The identifier of the provisioning artifact.}
}
\description{
Deletes the specified provisioning artifact (also known as a version) for the specified product.
}
\details{
You cannot delete a provisioning artifact associated with a product that was shared with you. You cannot delete the last provisioning artifact for a product, because a product must have at least one provisioning artifact.
}
\section{Accepted Parameters}{
\preformatted{delete_provisioning_artifact(
  AcceptLanguage = "string",
  ProductId = "string",
  ProvisioningArtifactId = "string"
)
}
}

