% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{delete_service_action}
\alias{delete_service_action}
\title{Deletes a self-service action}
\usage{
delete_service_action(Id, AcceptLanguage = NULL)
}
\arguments{
\item{Id}{[required] The self-service action identifier. For example, \code{act-fs7abcd89wxyz}.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Deletes a self-service action.
}
\section{Accepted Parameters}{
\preformatted{delete_service_action(
  Id = "string",
  AcceptLanguage = "string"
)
}
}

