% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_constraint}
\alias{describe_constraint}
\title{Gets information about the specified constraint}
\usage{
describe_constraint(Id, AcceptLanguage = NULL)
}
\arguments{
\item{Id}{The identifier of the constraint.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Gets information about the specified constraint.
}
\examples{

}
