% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_copy_product_status}
\alias{describe_copy_product_status}
\title{Gets the status of the specified copy product operation}
\usage{
describe_copy_product_status(CopyProductToken, AcceptLanguage = NULL)
}
\arguments{
\item{CopyProductToken}{The token for the copy product operation. This token is returned by CopyProduct.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Gets the status of the specified copy product operation.
}
\examples{

}
