% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_portfolio_share_status}
\alias{describe_portfolio_share_status}
\title{Gets the status of the specified portfolio share operation}
\usage{
describe_portfolio_share_status(
  PortfolioShareToken = "string"
)
}
\arguments{
\item{PortfolioShareToken}{[required] The token for the portfolio share operation. This token is returned either by CreatePortfolioShare or by DeletePortfolioShare.}
}
\description{
Gets the status of the specified portfolio share operation. This API can only be called by the master account in the organization.
}
\examples{

}
