% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_provisioning_artifact}
\alias{describe_provisioning_artifact}
\title{Gets information about the specified provisioning artifact (also known as a version) for the specified product}
\usage{
describe_provisioning_artifact(ProvisioningArtifactId, ProductId,
  AcceptLanguage = NULL, Verbose = NULL)
}
\arguments{
\item{ProvisioningArtifactId}{The identifier of the provisioning artifact.}

\item{ProductId}{The product identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Verbose}{Indicates whether a verbose level of detail is enabled.}
}
\description{
Gets information about the specified provisioning artifact (also known as a version) for the specified product.
}
\examples{

}
