% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_record}
\alias{describe_record}
\title{Gets information about the specified request operation}
\usage{
describe_record(AcceptLanguage = NULL, Id, PageToken = NULL,
  PageSize = NULL)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The record identifier of the provisioned product. This identifier is returned by the request operation.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of results, use null.}

\item{PageSize}{The maximum number of items to return with this call.}
}
\description{
Gets information about the specified request operation.
}
\details{
Use this operation after calling a request operation (for example, ProvisionProduct, TerminateProvisionedProduct, or UpdateProvisionedProduct).
}
\section{Accepted Parameters}{
\preformatted{describe_record(
  AcceptLanguage = "string",
  Id = "string",
  PageToken = "string",
  PageSize = 123
)
}
}

