% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_service_action}
\alias{describe_service_action}
\title{Describes a self-service action}
\usage{
describe_service_action(
  Id = "string",
  AcceptLanguage = "string"
)
}
\arguments{
\item{Id}{[required] The self-service action identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Describes a self-service action.
}
\examples{

}
