% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_tag_option}
\alias{describe_tag_option}
\title{Gets information about the specified TagOption}
\usage{
describe_tag_option(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] The TagOption identifier.}
}
\description{
Gets information about the specified TagOption.
}
\examples{

}
