% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{disable_aws_organizations_access}
\alias{disable_aws_organizations_access}
\title{Disable portfolio sharing through AWS Organizations feature}
\usage{
disable_aws_organizations_access()
}
\description{
Disable portfolio sharing through AWS Organizations feature. This feature will not delete your current shares but it will prevent you from creating new shares throughout your organization. Current shares will not be in sync with your organization structure if it changes after calling this API. This API can only be called by the master account in the organization.
}
\section{Accepted Parameters}{
\preformatted{disable_aws_organizations_access()
}
}

