% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{disassociate_principal_from_portfolio}
\alias{disassociate_principal_from_portfolio}
\title{Disassociates a previously associated principal ARN from a specified portfolio}
\usage{
disassociate_principal_from_portfolio(PortfolioId, PrincipalARN,
  AcceptLanguage = NULL)
}
\arguments{
\item{PortfolioId}{The portfolio identifier.}

\item{PrincipalARN}{The ARN of the principal (IAM user, role, or group).}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Disassociates a previously associated principal ARN from a specified portfolio.
}
\examples{

}
