% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{disassociate_tag_option_from_resource}
\alias{disassociate_tag_option_from_resource}
\title{Disassociates the specified TagOption from the specified resource}
\usage{
disassociate_tag_option_from_resource(
  ResourceId = "string",
  TagOptionId = "string"
)
}
\arguments{
\item{ResourceId}{[required] The resource identifier.}

\item{TagOptionId}{[required] The TagOption identifier.}
}
\description{
Disassociates the specified TagOption from the specified resource.
}
\examples{

}
