% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{enable_aws_organizations_access}
\alias{enable_aws_organizations_access}
\title{Enable portfolio sharing feature through AWS Organizations}
\usage{
enable_aws_organizations_access()
}
\description{
Enable portfolio sharing feature through AWS Organizations. This API will allow Service Catalog to receive updates on your organization in order to sync your shares with the current structure. This API can only be called by the master account in the organization.
}
\details{
By calling this API Service Catalog will use FAS credentials to call organizations:EnableAWSServiceAccess so that your shares can be in sync with any changes in your AWS Organizations.
}
\section{Accepted Parameters}{
\preformatted{enable_aws_organizations_access()
}
}

