% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{get_aws_organizations_access_status}
\alias{get_aws_organizations_access_status}
\title{Get the Access Status for AWS Organization portfolio share feature}
\usage{
get_aws_organizations_access_status()
}
\description{
Get the Access Status for AWS Organization portfolio share feature. This API can only be called by the master account in the organization.
}
\section{Accepted Parameters}{
\preformatted{get_aws_organizations_access_status()
}
}

