% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{list_constraints_for_portfolio}
\alias{list_constraints_for_portfolio}
\title{Lists the constraints for the specified portfolio and product}
\usage{
list_constraints_for_portfolio(
  AcceptLanguage = "string",
  PortfolioId = "string",
  ProductId = "string",
  PageSize = 123,
  PageToken = "string"
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{ProductId}{The product identifier.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of results, use null.}
}
\description{
Lists the constraints for the specified portfolio and product.
}
\examples{

}
