% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{list_portfolio_access}
\alias{list_portfolio_access}
\title{Lists the account IDs that have access to the specified portfolio}
\usage{
list_portfolio_access(PortfolioId, AcceptLanguage = NULL)
}
\arguments{
\item{PortfolioId}{The portfolio identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Lists the account IDs that have access to the specified portfolio.
}
\examples{

}
