% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{list_provisioning_artifacts}
\alias{list_provisioning_artifacts}
\title{Lists all provisioning artifacts (also known as versions) for the specified product}
\usage{
list_provisioning_artifacts(
  AcceptLanguage = "string",
  ProductId = "string"
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProductId}{[required] The product identifier.}
}
\description{
Lists all provisioning artifacts (also known as versions) for the specified product.
}
\examples{

}
