% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{list_service_actions}
\alias{list_service_actions}
\title{Lists all self-service actions}
\usage{
list_service_actions(
  AcceptLanguage = "string",
  PageSize = 123,
  PageToken = "string"
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of results, use null.}
}
\description{
Lists all self-service actions.
}
\examples{

}
