% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{list_tag_options}
\alias{list_tag_options}
\title{Lists the specified TagOptions or all TagOptions}
\usage{
list_tag_options(Filters = NULL, PageSize = NULL, PageToken = NULL)
}
\arguments{
\item{Filters}{The search filters. If no search filters are specified, the output includes all TagOptions.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of results, use null.}
}
\description{
Lists the specified TagOptions or all TagOptions.
}
\section{Accepted Parameters}{
\preformatted{list_tag_options(
  Filters = list(
    Key = "string",
    Value = "string",
    Active = TRUE|FALSE
  ),
  PageSize = 123,
  PageToken = "string"
)
}
}

