% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_package.R
\docType{package}
\name{paws.servicecatalog-package}
\alias{paws.servicecatalog}
\alias{paws.servicecatalog-package}
\title{paws.servicecatalog: AWS Service Catalog}
\description{
[AWS Service
    Catalog](https://aws.amazon.com/servicecatalog/) enables organizations
    to create and manage catalogs of IT services that are approved for use
    on AWS. To get the most out of this documentation, you should be
    familiar with the terminology discussed in [AWS Service Catalog
    Concepts](http://docs.aws.amazon.com/servicecatalog/latest/adminguide/what-is_concepts.html).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
