% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{provision_product}
\alias{provision_product}
\title{Provisions the specified product}
\usage{
provision_product(AcceptLanguage = NULL, ProductId,
  ProvisioningArtifactId, PathId = NULL, ProvisionedProductName,
  ProvisioningParameters = NULL, Tags = NULL,
  NotificationArns = NULL, ProvisionToken)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProductId}{[required] The product identifier.}

\item{ProvisioningArtifactId}{[required] The identifier of the provisioning artifact.}

\item{PathId}{The path identifier of the product. This value is optional if the product has a default path, and required if the product has more than one path. To list the paths for a product, use ListLaunchPaths.}

\item{ProvisionedProductName}{[required] A user-friendly name for the provisioned product. This value must be unique for the AWS account and cannot be updated after the product is provisioned.}

\item{ProvisioningParameters}{Parameters specified by the administrator that are required for provisioning the product.}

\item{Tags}{One or more tags.}

\item{NotificationArns}{Passed to CloudFormation. The SNS topic ARNs to which to publish stack-related events.}

\item{ProvisionToken}{[required] An idempotency token that uniquely identifies the provisioning request.}
}
\description{
Provisions the specified product.
}
\details{
A provisioned product is a resourced instance of a product. For example, provisioning a product based on a CloudFormation template launches a CloudFormation stack and its underlying resources. You can check the status of this request using DescribeRecord.

If the request contains a tag key with an empty list of values, there is a tag conflict for that key. Do not include conflicted keys as tags, or this causes the error "Parameter validation failed: Missing required parameter in Tags[\emph{N}]:\emph{Value}".
}
\section{Accepted Parameters}{
\preformatted{provision_product(
  AcceptLanguage = "string",
  ProductId = "string",
  ProvisioningArtifactId = "string",
  PathId = "string",
  ProvisionedProductName = "string",
  ProvisioningParameters = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  NotificationArns = list(
    "string"
  ),
  ProvisionToken = "string"
)
}
}

