% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{reject_portfolio_share}
\alias{reject_portfolio_share}
\title{Rejects an offer to share the specified portfolio}
\usage{
reject_portfolio_share(PortfolioId, AcceptLanguage = NULL)
}
\arguments{
\item{PortfolioId}{The portfolio identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Rejects an offer to share the specified portfolio.
}
\examples{

}
