% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{search_products_as_admin}
\alias{search_products_as_admin}
\title{Gets information about the products for the specified portfolio or all products}
\usage{
search_products_as_admin(
  AcceptLanguage = "string",
  PortfolioId = "string",
  Filters = list(
    list(
      "string"
    )
  ),
  SortBy = "Title"|"VersionCount"|"CreationDate",
  SortOrder = "ASCENDING"|"DESCENDING",
  PageToken = "string",
  PageSize = 123,
  ProductSource = "ACCOUNT"
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{The portfolio identifier.}

\item{Filters}{The search filters. If no search filters are specified, the output includes all products to which the administrator has access.}

\item{SortBy}{The sort field. If no value is specified, the results are not sorted.}

\item{SortOrder}{The sort order. If no value is specified, the results are not sorted.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of results, use null.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{ProductSource}{Access level of the source of the product.}
}
\description{
Gets information about the products for the specified portfolio or all products.
}
\examples{

}
