% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{update_constraint}
\alias{update_constraint}
\title{Updates the specified constraint}
\usage{
update_constraint(Id, AcceptLanguage = NULL, Description = NULL)
}
\arguments{
\item{Id}{The identifier of the constraint.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Description}{The updated description of the constraint.}
}
\description{
Updates the specified constraint.
}
\examples{

}
