% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{update_portfolio}
\alias{update_portfolio}
\title{Updates the specified portfolio}
\usage{
update_portfolio(
  AcceptLanguage = "string",
  Id = "string",
  DisplayName = "string",
  Description = "string",
  ProviderName = "string",
  AddTags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  RemoveTags = list(
    "string"
  )
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The portfolio identifier.}

\item{DisplayName}{The name to use for display purposes.}

\item{Description}{The updated description of the portfolio.}

\item{ProviderName}{The updated name of the portfolio provider.}

\item{AddTags}{The tags to add.}

\item{RemoveTags}{The tags to remove.}
}
\description{
Updates the specified portfolio.
}
\details{
You cannot update a product that was shared with you.
}
\examples{

}
