% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{update_provisioned_product}
\alias{update_provisioned_product}
\title{Requests updates to the configuration of the specified provisioned product}
\usage{
update_provisioned_product(UpdateToken, AcceptLanguage = NULL,
  ProvisionedProductName = NULL, ProvisionedProductId = NULL,
  ProductId = NULL, ProvisioningArtifactId = NULL, PathId = NULL,
  ProvisioningParameters = NULL)
}
\arguments{
\item{UpdateToken}{The idempotency token that uniquely identifies the provisioning update request.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProvisionedProductName}{The updated name of the provisioned product. You cannot specify both \code{ProvisionedProductName} and \code{ProvisionedProductId}.}

\item{ProvisionedProductId}{The identifier of the provisioned product. You cannot specify both \code{ProvisionedProductName} and \code{ProvisionedProductId}.}

\item{ProductId}{The identifier of the provisioned product.}

\item{ProvisioningArtifactId}{The identifier of the provisioning artifact.}

\item{PathId}{The new path identifier. This value is optional if the product has a default path, and required if the product has more than one path.}

\item{ProvisioningParameters}{The new parameters.}
}
\description{
Requests updates to the configuration of the specified provisioned product.
}
\details{
If there are tags associated with the object, they cannot be updated or added. Depending on the specific updates requested, this operation can update with no interruption, with some interruption, or replace the provisioned product entirely.

You can check the status of this request using DescribeRecord.
}
\examples{

}
