% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{update_provisioning_artifact}
\alias{update_provisioning_artifact}
\title{Updates the specified provisioning artifact (also known as a version) for the specified product}
\usage{
update_provisioning_artifact(ProductId, ProvisioningArtifactId,
  AcceptLanguage = NULL, Name = NULL, Description = NULL, Active = NULL)
}
\arguments{
\item{ProductId}{The product identifier.}

\item{ProvisioningArtifactId}{The identifier of the provisioning artifact.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Name}{The updated name of the provisioning artifact.}

\item{Description}{The updated description of the provisioning artifact.}

\item{Active}{Indicates whether the product version is active.}
}
\description{
Updates the specified provisioning artifact (also known as a version) for the specified product.
}
\details{
You cannot update a provisioning artifact for a product that was shared with you.
}
\examples{

}
