% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{update_service_action}
\alias{update_service_action}
\title{Updates a self-service action}
\usage{
update_service_action(Id, Name = NULL, Definition = NULL,
  Description = NULL, AcceptLanguage = NULL)
}
\arguments{
\item{Id}{[required] The self-service action identifier.}

\item{Name}{The self-service action name.}

\item{Definition}{A map that defines the self-service action.}

\item{Description}{The self-service action description.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Updates a self-service action.
}
\section{Accepted Parameters}{
\preformatted{update_service_action(
  Id = "string",
  Name = "string",
  Definition = list(
    "string"
  ),
  Description = "string",
  AcceptLanguage = "string"
)
}
}

