% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{accept_portfolio_share}
\alias{accept_portfolio_share}
\title{Accepts an offer to share the specified portfolio}
\usage{
accept_portfolio_share(PortfolioId, AcceptLanguage = NULL)
}
\arguments{
\item{PortfolioId}{The portfolio identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Accepts an offer to share the specified portfolio.
}
\examples{

}
