% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{associate_principal_with_portfolio}
\alias{associate_principal_with_portfolio}
\title{Associates the specified principal ARN with the specified portfolio}
\usage{
associate_principal_with_portfolio(PortfolioId, PrincipalARN, PrincipalType,
  AcceptLanguage = NULL)
}
\arguments{
\item{PortfolioId}{The portfolio identifier.}

\item{PrincipalARN}{The ARN of the principal (IAM user, role, or group).}

\item{PrincipalType}{The principal type. The supported value is \code{IAM}.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Associates the specified principal ARN with the specified portfolio.
}
\examples{

}
