% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{associate_product_with_portfolio}
\alias{associate_product_with_portfolio}
\title{Associates the specified product with the specified portfolio}
\usage{
associate_product_with_portfolio(
  AcceptLanguage = "string",
  ProductId = "string",
  PortfolioId = "string",
  SourcePortfolioId = "string"
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProductId}{[required] The product identifier.}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{SourcePortfolioId}{The identifier of the source portfolio.}
}
\description{
Associates the specified product with the specified portfolio.
}
\examples{

}
