% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{create_portfolio}
\alias{create_portfolio}
\title{Creates a portfolio}
\usage{
create_portfolio(
  AcceptLanguage = "string",
  DisplayName = "string",
  Description = "string",
  ProviderName = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  IdempotencyToken = "string"
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{DisplayName}{[required] The name to use for display purposes.}

\item{Description}{The description of the portfolio.}

\item{ProviderName}{[required] The name of the portfolio provider.}

\item{Tags}{One or more tags.}

\item{IdempotencyToken}{[required] A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.}
}
\description{
Creates a portfolio.
}
\examples{

}
