% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{create_product}
\alias{create_product}
\title{Creates a product}
\usage{
create_product(Name, Owner, ProductType, ProvisioningArtifactParameters,
  IdempotencyToken, AcceptLanguage = NULL, Description = NULL,
  Distributor = NULL, SupportDescription = NULL, SupportEmail = NULL,
  SupportUrl = NULL, Tags = NULL)
}
\arguments{
\item{Name}{The name of the product.}

\item{Owner}{The owner of the product.}

\item{ProductType}{The type of product.}

\item{ProvisioningArtifactParameters}{The configuration of the provisioning artifact.}

\item{IdempotencyToken}{A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Description}{The description of the product.}

\item{Distributor}{The distributor of the product.}

\item{SupportDescription}{The support information about the product.}

\item{SupportEmail}{The contact email for product support.}

\item{SupportUrl}{The contact URL for product support.}

\item{Tags}{One or more tags.}
}
\description{
Creates a product.
}
\examples{

}
