% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_constraint}
\alias{describe_constraint}
\title{Gets information about the specified constraint}
\usage{
describe_constraint(AcceptLanguage = NULL, Id)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The identifier of the constraint.}
}
\description{
Gets information about the specified constraint.
}
\section{Accepted Parameters}{
\preformatted{describe_constraint(
  AcceptLanguage = "string",
  Id = "string"
)
}
}

