% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_portfolio}
\alias{describe_portfolio}
\title{Gets information about the specified portfolio}
\usage{
describe_portfolio(AcceptLanguage = NULL, Id)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The portfolio identifier.}
}
\description{
Gets information about the specified portfolio.
}
\section{Accepted Parameters}{
\preformatted{describe_portfolio(
  AcceptLanguage = "string",
  Id = "string"
)
}
}

