% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_product_as_admin}
\alias{describe_product_as_admin}
\title{Gets information about the specified product}
\usage{
describe_product_as_admin(
  AcceptLanguage = "string",
  Id = "string"
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The product identifier.}
}
\description{
Gets information about the specified product. This operation is run with administrator access.
}
\examples{

}
