% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_provisioned_product_plan}
\alias{describe_provisioned_product_plan}
\title{Gets information about the resource changes for the specified plan}
\usage{
describe_provisioned_product_plan(AcceptLanguage = NULL, PlanId,
  PageSize = NULL, PageToken = NULL)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PlanId}{[required] The plan identifier.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of results, use null.}
}
\description{
Gets information about the resource changes for the specified plan.
}
\section{Accepted Parameters}{
\preformatted{describe_provisioned_product_plan(
  AcceptLanguage = "string",
  PlanId = "string",
  PageSize = 123,
  PageToken = "string"
)
}
}

