% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{describe_provisioning_artifact}
\alias{describe_provisioning_artifact}
\title{Gets information about the specified provisioning artifact (also known as a version) for the specified product}
\usage{
describe_provisioning_artifact(AcceptLanguage = NULL,
  ProvisioningArtifactId, ProductId, Verbose = NULL)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProvisioningArtifactId}{[required] The identifier of the provisioning artifact.}

\item{ProductId}{[required] The product identifier.}

\item{Verbose}{Indicates whether a verbose level of detail is enabled.}
}
\description{
Gets information about the specified provisioning artifact (also known as a version) for the specified product.
}
\section{Accepted Parameters}{
\preformatted{describe_provisioning_artifact(
  AcceptLanguage = "string",
  ProvisioningArtifactId = "string",
  ProductId = "string",
  Verbose = TRUE|FALSE
)
}
}

