% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{disassociate_principal_from_portfolio}
\alias{disassociate_principal_from_portfolio}
\title{Disassociates a previously associated principal ARN from a specified portfolio}
\usage{
disassociate_principal_from_portfolio(
  AcceptLanguage = "string",
  PortfolioId = "string",
  PrincipalARN = "string"
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{PrincipalARN}{[required] The ARN of the principal (IAM user, role, or group).}
}
\description{
Disassociates a previously associated principal ARN from a specified portfolio.
}
\examples{

}
