% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{list_portfolios_for_product}
\alias{list_portfolios_for_product}
\title{Lists all portfolios that the specified product is associated with}
\usage{
list_portfolios_for_product(ProductId, AcceptLanguage = NULL,
  PageToken = NULL, PageSize = NULL)
}
\arguments{
\item{ProductId}{The product identifier.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of results, use null.}

\item{PageSize}{The maximum number of items to return with this call.}
}
\description{
Lists all portfolios that the specified product is associated with.
}
\examples{

}
