% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{list_principals_for_portfolio}
\alias{list_principals_for_portfolio}
\title{Lists all principal ARNs associated with the specified portfolio}
\usage{
list_principals_for_portfolio(AcceptLanguage = NULL, PortfolioId,
  PageSize = NULL, PageToken = NULL)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of results, use null.}
}
\description{
Lists all principal ARNs associated with the specified portfolio.
}
\section{Accepted Parameters}{
\preformatted{list_principals_for_portfolio(
  AcceptLanguage = "string",
  PortfolioId = "string",
  PageSize = 123,
  PageToken = "string"
)
}
}

