% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicecatalog_operations.R
\name{list_provisioning_artifacts_for_service_action}
\alias{list_provisioning_artifacts_for_service_action}
\title{Lists all provisioning artifacts (also known as versions) for the specified self-service action}
\usage{
list_provisioning_artifacts_for_service_action(ServiceActionId,
  PageSize = NULL, PageToken = NULL, AcceptLanguage = NULL)
}
\arguments{
\item{ServiceActionId}{The self-service action identifier. For example, \code{act-fs7abcd89wxyz}.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of results, use null.}

\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}
}
\description{
Lists all provisioning artifacts (also known as versions) for the specified self-service action.
}
\examples{

}
